==> prep/bd/spp/v2020/1_iucn_spp_lists.Rmd <==
---
title: 'Species Sub-goal: IUCN Species List'
subtitle: 'OHI+ Palmyra 2020 Assessment'
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: show
    toc: true
    toc_depth: 3 
    toc_float: yes
    number_sections: false
    theme: cerulean
    highlight: haddock
    includes: 
     in_header: '~/github/pal-prep/src/templates/pal_hdr.html' 
  pdf_document:
    toc: true
---

# Summary

This script is the first step toward calculating the species status. It takes the OHI Palmyra region shapefile and compares it to the global shapefiles we have from IUCN to extract all species maps that fall within 50 nm of Palmyra Atoll. We then get the conservation status for all these species. A lot of this code dips into the `spp_risk_dists` repo from Casey O'Hara: 
https://github.com/saraorofino/spp_risk_dists  

# Data 

Species maps and conservation status were pulled from the IUCN Red List.  

Citation: IUCN 2020. IUCN Red List of Threatened Species. Version 2020-1 <www.iucnredlist.org>   
  * __Accessed__: 9 June 2020   
  * __Time range__: 1965-2020 (discrete past assessments by species)   
  * __Format__:  JSON   

# Setup

```{r setup, message=F, warning=F, results='hide'}

library(tidyverse)
library(sf)
library(raster)
library(rfishbase)
library(rredlist)
library(taxize)
library(rentrez) # To use the api key for the NCBI database
library(here)

# Source and set file paths
source(here('src/R/common.R'))

dir_goal   <- '~/github/pal-prep/prep/bd/spp/v2020'

# Palmyra region shapefile 50nm boundary
pal_shp <- pal_monument

# api keys for IUCN 
api_file <- file.path(dir_M, 'git-annex/globalprep/spp', 
                      'api_key_gc.csv')
api_key <- scan(api_file, what = 'character')
```

# Methods

## Get Cell IDs and Species List

First step is to get a list of the cell IDs contained in the pal_region

```{r global-cells, eval=F}
# These are global cell ids in raster format pulled from the spp_risk_dists repo.
cells <- raster("~/github/spp_risk_dists/_spatial/cell_id_rast.tif")
```

  
Then I select only those cells in Palmyra and use that list to query the larger species data. First we transform our regions shapefile into the same CRS as the cell ID raster (`cells`).

```{r reproject-rgns, eval=F}

# Reproject to our CRS
pal_reproj <- st_transform(pal_shp,
              crs = "+proj=cea +lon_0=0 +lat_ts=45 +x_0=0 +y_0=0 +ellps=WGS84 +units=m +no_defs")

# Plot to check
plot(cells)
plot(pal_reproj, add = T)
```

Extract the cell ids for our region 

```{r extract-pal-cell-ids, eval=F}
pal_cells <- raster::extract(cells, pal_reproj) %>%
  unlist()
```

Read in files that contain all the species ranges and information

```{r spp-files, eval=F}
# .csv file that lists all species and the file path to their map
spp_maps <- read_csv('~/github/spp_risk_dists/_data/spp_marine_maps_2019-2.csv',
                     col_types = 'ddciccc')

# file with species information to link to iucn_sid at the end
spp_info <- read_csv("/home/ohara/git-annex/spp_risk_dists/iucn/spp_info_from_api_2019-2.csv")
```

The following forloop goes through each species:  
- finds it's species path   
- reads in the species .csv that lists all global cells where it is found   
- filters that list to just those in Palmyra   
- returns an aggregated dataframe (`taxa_cells_df`) with all species scientific names, their unique ids (sids), and cellIDs   

```{r get-spp-files, eval=F}

# Grab each taxa's folder filepath
taxa <- spp_maps$dbf_file %>%
    unique() %>%
    str_replace('\\....$', '')

# Create an empty list that is the length of all taxa  
taxa_cells_list <- vector('list', length = length(taxa))


# For each taxa, grab the species map (raster)
# filter to only keep those cells in Palmyra

for(i in seq_along(taxa)) { ### i <- 5
    taxon <- taxa[i]
    print(i)
    
    spp_ids_in_taxon <- spp_maps %>%
     filter(str_detect(dbf_file, taxon)) %>%
      .$iucn_sid
    cat(sprintf('processing %s spp in %s...\n', length(spp_ids_in_taxon), taxon)) 
    
    spp_cells <- parallel::mclapply(spp_ids_in_taxon, mc.cores = 32,
                                    FUN = function(x) { ###x <- spp_ids_in_taxon[1]
                                      f <- file.path('/home/ohara/git-annex/spp_risk_dists/spp_rasters_2019',
                                                     sprintf('iucn_sid_%s.csv', x))
                                      if(file.exists(f)) {
                                        y <- read_csv(f, col_types = 'di') %>%
                                          mutate(iucn_sid = x) %>%
                                          dplyr::select(-presence)  %>%
                                          filter(cell_id %in% pal_cells)
                                      } else {
                                        y <- data.frame(cell_id = NA,
                                                        iucn_sid = x, 
                                                        f = f, error = 'file not found')
                                      }
                                      return(y)
                                    }) %>%
      bind_rows() %>%
      mutate(spp_gp = taxon)
    
    taxa_cells_list[[i]] <- spp_cells
}
  
taxa_cells_df <- taxa_cells_list %>%
    bind_rows()  %>%
    filter(!is.na(cell_id)) %>%
    dplyr::select(iucn_sid) %>%
    distinct() %>%
    left_join(spp_info)

# Save raw output to the raw folder
write.csv(taxa_cells_df, file.path(dir_goal, 'raw/spp_list_raw.csv'))
```

## Add Common Names

Attach common names using `rredlist`

```{r get-common-name, eval=F}

# Read in raw data list:
spp_list_raw <- read_csv(file.path(dir_goal, 'raw/spp_list_raw.csv')) %>% 
  dplyr::select(-X1)

# Create list of scientific names of species
scinames <- spp_list_raw$sciname

# Create loop around redlist function rl_search to find common names  
# Create empty data frame that will be filled
iucn_common_names <- data.frame() 


for(i in 1:length(scinames)){
  sp <- scinames[i]
  print(i)

  comm <- rredlist::rl_search(sp, key = api_key)$result$main_common_name
  
  df <- data.frame(sciname = sp,
                   common = comm)
  
  iucn_common_names <- rbind(iucn_common_names, df) %>%
    mutate(common = as.character(common),
         sciname = as.character(sciname))
}
```

Check out the ones we are still missing for common name   

```{r missing-spp, eval=F}

miss_sp <- iucn_common_names %>%
  filter(is.na(common))

```

We are still missing 106 common names. Check fishbase

```{r fishbase, eval=F}

# Select scientific names of missing species
sp <- iucn_common_names %>%
  filter(is.na(common))  %>%
  dplyr::select(sciname)

# Check fishbase for common names
fb_list <- species(sp$sciname) %>%
  dplyr::select(Species, FBname)

# Filter out NAs
fb <- fb_list %>% 
  filter(!is.na(FBname)) 

```

This only got us two more common names we need to add back in.  

```{r add-fishbase, eval=F}

# Add back in common names
out <- iucn_common_names %>%
  left_join(fb_list, by = c("sciname" = "Species")) %>%
  mutate(common_fb = ifelse(is.na(common), FBname, common)) %>%
  dplyr::select(-FBname, -common)

```

Use the `taxize` R package to try and get more of the missing common names

```{r taxize, eval=F}

# Missing scientific names
missing_scinames <- out %>%
  filter(is.na(common_fb)) %>%
  .$sciname

# Pull out common names from taxize
common_names <- c()
for(i in 1:length(missing_scinames)){ #i <- 1
  sp <- missing_scinames[i]
  comm <- taxize::sci2comm(sp, db = "itis")
  print(comm)
  
common_names <- c(common_names, comm)
}


# Turn list into dataframe
df <- do.call(rbind,lapply(common_names,data.frame))
df$sciname <- rownames(df)
df2 <- df %>%
  rename(common = X..i..) %>%
  filter(!is.na(common)) %>%
  mutate(sciname = str_extract(sciname, "[^.]+")) # Remove numbers

## Try using the NCBI in taxize
uids <- get_uid(missing_scinames)

# Combine uids with sciname 
uids_found <- data.frame(sciname = missing_scinames, uid = uids)
uids_found$sciname <- as.character(uids_found$sciname)

# Get an api key using use_entrez with the account I created through NCBI 
# Set api key in the environment using rentrez::set_entrez_key()
taxize::use_entrez()
rentrez::set_entrez_key("c6d9419d922a77ebe9e15f9a4c0e46a12108")

# Search for common names using the NCBI database
common_names2 <- sci2comm(uids_found$sciname, db = 'ncbi') # only found four more 
common_found <- common_names2 %>% 
  unlist() %>% 
  as.data.frame() %>% 
  rownames_to_column(., "sciname") %>% 
  rename("ncbi_common" = ".")

# Make sure names are characters not factors 
common_found$ncbi_common <- as.character(common_found$ncbi_common)

```

None of the 104 remaining species had common names in the taxize 'itis' database but I found four more in the 'ncbi' database

## Select Species Information

Add common names and select final list of spp information

```{r add-common-names, eval=F}

# Add back in common names from fishbase and ncbi 
add_common <- iucn_common_names %>%
  left_join(common_found, by = "sciname") %>%
  mutate(common = ifelse(is.na(common), ncbi_common, common)) %>% 
  dplyr::select(-ncbi_common) %>% 
  left_join(out, by = "sciname") %>% 
  mutate(common = ifelse(is.na(common), common_fb, common)) %>% 
  dplyr::select(-common_fb)

iucn_spp_info <- spp_list_raw %>%
  dplyr::select(iucn_sid, sciname, class, family, population, category) %>% 
  left_join(add_common, by = "sciname") %>% 
  mutate(class = tolower(class), family = tolower(family)) %>% 
  distinct()

## Experiment with taxonomic groupings:
spp_class <- iucn_spp_info %>% 
  group_by(class) %>% 
  tally() # This gives me 10 classes of species - seems reasonable

# Common names/animals in the different classes
class_common <- c("ray finned fishes", "corals/sea anemones", "birds",
                  "sharks/rays", "snails/slugs", "sea cucumbers",
                  "crabs/lobsters", "mammals", "hagfish", "reptiles") 
spp_class$common <- class_common

spp_family <- iucn_spp_info %>% 
  group_by(family) %>% 
  tally() # 51 families - seems like too many

## Include the class in the output for grouping later, remove family
iucn_spp_info <- iucn_spp_info %>% 
  dplyr::select(-family)


# Write output to int folder
write.csv(iucn_spp_info, file.path(dir_goal, 'int/iucn_spp_info.csv'))
```

## Explore Missing Common Names

Some exploration of the species missing common names:  

```{r missing_spp_explore, eval = F}

scinames_explore <- miss_sp %>%
  separate(sciname, sep = " ", into = c("genus", "family")) %>% 
  group_by(genus) %>% 
  summarize(
    count = length(family))

```

Looks like there are 20 genus' in the missing data, 9 of them have more than one species in the genus:   

 - `Acopora` are small polyp stony corals (5 species) 
 - `Conus` are types of sea snails (43 species)   
 - `Favia` are reef-building stony corals (4 species)  
 - `Favites` are also stony corals (5 species)   
 - `Holothuria` are sea cucumbers (15 species)     
 - `Leptastrea` are also reef-building stony corals (4 species)
 - `Montastrea` are colonial stony corals (2 species)       
 - `Montipora` are Scleractinian corals (12 species)     
 - `Platygyra` are also stony corals (5 species)
