==> prep/bd/spp/v2020/3_spp_status.Rmd <==
---
title: 'Species Sub-goal: Status'
subtitle: 'OHI+ Palmyra 2020 Assessment'
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: show
    toc: true
    toc_depth: 3 
    toc_float: yes
    number_sections: false
    theme: cerulean
    highlight: haddock
    includes: 
     in_header: '~/github/pal-prep/src/templates/pal_hdr.html' 
  pdf_document:
    toc: true
---

# Summary

This is the third script in the species data prep. It takes the fully combined IUCN and Palmyra species lists and calculates the status score using weights for each species' IUCN conservation status.   

# Data 

Data used in this script is the result of processing species lists in `01_iucn_spp_lists.Rmd` and `02_pal_spp_lits.Rmd`  

# Setup 

```{r setup, message=F, warning=F, results='hide'}

library(tidyverse)
library(here)

# Source and set file paths
source(here('src/R/common.R'))

dir_goal   <- '~/github/pal-prep/prep/bd/spp/v2020'
```

# Methods

## Calculate Status 

**Reference Point:** All species have an IUCN status of "Least Concern"   

A 0-1 extinction risk score is assigned to each species based on IUCN conservation status:   
 
  * 'LC' = 0, 'NT' = 0.2, 'CD' = 0.3, 'VU' = 0.4, 'T' = 0.6, 'EN' = 0.6, 'CR' = 0.8, 'EX' = 1.0  

The status is first calculated for each class by averaging the risk scores. Then, the risk status is averaged over all the classes. The overall score is rescaled so a risk factor of 0.75 is equal to zero. This assigns a score of 0 when 75% species are extinct, a level comparable to the five documented mass extinctions (Barnosky et al. 2011) and would constitute a catastrophic loss of biodiversity.    


```{r calc-spp-status, eval=F}
# Use the combined all species lists
all_spp <- read_csv(file.path(dir_goal, 'int/all_spp_list.csv')) 


# Create category weights that will be used to calculate scores
pop_cat <- data.frame(category  = c("LC", "NT", "VU", "EN", "CR", "EX", "T", "CD", "NE", "DD"), 
                      cat_score = c(   0,  0.2,  0.4,  0.6,  0.8,  1.0, 0.6,  0.3,   NA,  NA),
                      stringsAsFactors = FALSE)

# Join category weights to species list
spp_status <- all_spp %>% 
  left_join(pop_cat, by = "category") %>% 
  filter(!is.na(cat_score)) %>%
  distinct() %>% 
  mutate(region_id = 1,
         status = cat_score,
         year = 2019) %>%
  dplyr::select(region_id, sciname, class, common, status, year) 

# Save to output folder
write_csv(spp_status, file.path(dir_goal, "output/spp_status.csv"))
```


## Save to Toolbox 

```{r save-to-toolbox, eval=F}
spp_status <- read_csv(file.path(dir_goal, "output/spp_status.csv"))
write_csv(spp_status, file.path(dir_scores, 'layers/spp_status.csv'))
```
