==> prep/bd/spp/v2020/4_spp_trend.Rmd <==
---
title: 'Species Sub-goal: Trend'
subtitle: 'OHI+ Palmyra 2020 Assessment'
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: show
    toc: true
    toc_depth: 3 
    toc_float: yes
    number_sections: false
    theme: cerulean
    highlight: haddock
    includes: 
     in_header: '~/github/pal-prep/src/templates/pal_hdr.html' 
  pdf_document:
    toc: true 
---

# Summary 

This script estimates the population trend in species over time, using historic assessments and linear regression. Finally, missing species trends are gapfilled using the IUCN population trend designations of Increasing, Decreasing, and Stable.      

# Data 

Historic assessments are based on the IUCN Red List past assessments.   


Citation: IUCN 2020. IUCN Red List of Threatened Species. Version 2020-1 <www.iucnredlist.org>     
  * __Accessed__: 9 June 2020    
  * __Time range__: 1965-2020 (discrete past assessments by species)   
  * __Format__:  JSON     

# Setup

```{r setup, eval=F, messsage=F, warning=F, results='hide'}

library(tidyverse)
library(rredlist)
library(here)

# Source and set file paths
source(here('src/R/common.R'))

dir_goal   <- '~/github/pal-prep/prep/bd/spp/v2020'

# api keys for IUCN 
api_file    <- file.path(dir_M, 'git-annex/globalprep/spp', 
                      'api_key_gc.csv')
api_key     <- scan(api_file, what = 'character')
api_version <- '2020-1'
```

# Methods

## Get Historic Threat 

The `rl_history()` function retrieves all previous assessments and status for each species.

```{r get-historical-status, eval = F}

# Use all species list
spp_list <- read_csv(file.path(dir_goal, 'int/all_spp_list.csv'))

df_iucn <- data.frame()

for(i in 1:nrow(spp_list)){
  ###i <- 1
  print(i)
  
  sp <- as.character(spp_list[i,2]) # Grab scientific name
  
  possibleError <- tryCatch(
   
    sp_history <- rl_history(sp, key = api_key)$result %>%
                  mutate(sciname = sp),
      
    error=function(e) e)
  
  if(inherits(possibleError, "error")) next
    
  df_iucn <- rbind(df_iucn, sp_history)
}

# Save to int because this loop takes awhile to run 
write_csv(df_iucn, file.path(dir_goal, "int/raw_spp_assess.csv"))
```

## Get Population Trend from IUCN

Using `rl_search` find the population trend from the IUCN for each species. Trends will be increasing, decreasing, stable or NA   

```{r get-pop-trend, eval=F}

spp_list <- read_csv(file.path(dir_goal, "int/all_spp_list.csv")) 

# Loop for each species to grab population trend
df_iucn_trend <- data.frame()

for(i in 1:nrow(spp_list)){
  
  print(i)
  sp <- as.character(spp_list[i,2]) # Grab scientific name
  
  tr <- rl_search(sp, key = api_key)$result$population_trend
  
  if(is.null(tr)){
  
  df2 <- data.frame(sciname = sp,
                    trend = NA)
  }else{
    df2 <- data.frame(sciname = sp,
                      trend = tr)
  }
  
  df_iucn_trend <- rbind(df_iucn_trend, df2)
}

# Check how many are unknown:
trend_unknown <- df_iucn_trend %>% 
  filter(trend == "Unknown") # 286 are unknown

# Save file to int folder:
write_csv(df_iucn_trend, file.path(dir_goal, 'int/iucn_pop_trend.csv'))
```


## Add Past Category Designations

Some categories may have changed designations in the past. A list of old categories was compiled by Casey O'Hara and pulled from the OHI Global Prep repository   

``` {r add-past-categories, eval=F}

spp_hist   <- read_csv(file.path(dir_goal, 'int/raw_spp_assess.csv'))
cat_lookup <- read_csv(file.path(dir_goal, "raw/risk_code_lookup.csv"))

# Join past categories    
spp_past_df <- spp_hist %>%
    left_join(cat_lookup, by = c('code', 'category')) %>%
    dplyr::rename(
           old_cat  = code,
           cat_txt  = category,
           cat_ts   = code_current) 

# Define category scores
pop_cat <- data.frame(cat_ts       = c('LC', 'NT', 'VU', 'EN', 'CR', 'EX', 'NE', 'DD'),
                        cat_ts_score = c( 0.0,  0.2,  0.4,  0.6,  0.8,  1.0,  NA,   NA)) 

# Add scores to past categories 
spp_past_df1 <- spp_past_df %>%
    left_join(pop_cat, by = 'cat_ts') %>%
    filter(!is.na(cat_ts_score) & !is.na(year)) %>%
    arrange(sciname, year) %>%
    dplyr::select(sciname, year, cat_ts, cat_ts_score) %>%
    mutate(iucn_version = api_version) # What year it is pulling from 
```

## Linear Model

Create data frames for modeling 

- `cat_trend` binds our historic pull to the trend, contains all info for all years
- `lm_trend` filters out only ones that have enough info to create a linear model on 

``` {r df-modeling, eval=F}

pop_trends <- read_csv(file.path(dir_goal, "int/iucn_pop_trend.csv")) %>%
  dplyr::rename(pop_trend = trend) # Read in population trends grabbed from IUCN

# Group historic trends and population trends
cat_trend <- pop_trends %>%
  left_join(spp_past_df1, by = "sciname") %>%
  group_by(sciname) %>%
  dplyr::mutate(n_assess = n()) %>%
  ungroup()

cat_trend$year <- as.numeric(cat_trend$year) 

# Run a linear regression to get trend based on historical assessments
lm_trend <- cat_trend %>%
  filter(n_assess >= 2) %>% # Only include ones with enough assessments
  group_by(sciname) %>%
  do(calc_trend = lm(cat_ts_score ~ year, data = .)[['coefficients']][['year']]) %>%
  mutate(calc_trend = as.numeric(calc_trend)) %>%
  mutate(calc_trend = round(calc_trend, 5)) 
```

## Compare Trends

Compare IUCN trends to the model trends   

``` {r compare-trends, eval=F}

# Combine the lm trends with the iucn population trends and compare
comp_trend <- cat_trend %>%
  dplyr::filter(n_assess > 1) %>%
  dplyr::group_by(sciname) %>%
  dplyr::summarize(years  = paste(year, collapse = ', '),
            cat_ts = paste(cat_ts, collapse = ', '),
            scores = paste(cat_ts_score, collapse = ', '),
            pop_trend_iucn = first(pop_trend)) %>%
  ungroup() %>%
  left_join(lm_trend, by = c('sciname'))

comp_trend2 <- comp_trend %>%
  filter(!is.na(pop_trend_iucn)) %>%
  mutate(pop_trend = case_when(pop_trend_iucn == 'Decreasing' ~ -1,
                               pop_trend_iucn == 'Increasing' ~  1,
                               pop_trend_iucn == 'Stable'     ~  0,
                               TRUE                           ~ NA_real_)
         )
```

## Trend Regression

Take `calc_trend` and model it on the IUCN population trend. Here we're coding the trend in *population* not risk   

``` {r calc-pop-trend, eval=F}  

trend_regr   <- lm(calc_trend ~ pop_trend, data = comp_trend2)
trend_coeffs <- summary(trend_regr)

```

## Trend Scores

Calculate the trend scores based on the regression coefficients 

``` {r grab-values-lm, eval=F}

pop_trends <- read_csv(file.path(dir_goal, "int/iucn_pop_trend.csv")) %>%
  dplyr::rename(pop_trend = trend) # Read in population trends grabbed from IUCN

# Get values from linear regression:
slope <- trend_coeffs$coefficients[1] %>% round(5) # Grabs slope from model

# Gapfilling if the model didn't find anything
trend_score_lookup <- c('Decreasing' = -slope, 'Stable' = 0, 'Increasing' = slope) 

trend <- comp_trend2 %>%
  dplyr::select(sciname, calc_trend) %>%
  distinct()

# Join population trends and regression trends
spp_trends <- pop_trends %>%
  left_join(spp_past_df1, by = c('sciname')) %>%
  distinct() %>%
  left_join(trend, by = c('sciname')) %>%
  mutate(trend_score  = ifelse(is.na(calc_trend), trend_score_lookup[pop_trend], calc_trend),
         trend_source = case_when(!is.na(calc_trend)  ~ 'lm',
                                  !is.na(trend_score) ~ 'regr',
                                  TRUE ~ NA_character_)) %>%
  dplyr::select(sciname,  pop_trend, trend_score, trend_source) %>%
  distinct() 

# Check for trend score NAs:
trend_na <- spp_trends %>% 
  filter(is.na(trend_score)) # 292 NA values - gapfill using the IUCN population trend

# Save to int folder
write_csv(spp_trends, file.path(dir_goal, 'int/spp_trends_model.csv'))
```

## Gapfill Trend Scores

Gapfill the species without trend scores from the linear regression model. Using the IUCN population trend add scores based on the following scale:   

Increasing =  0.025   
Decreasing = -0.025   
Stable     =  0   
Unknown    =  NA

```{r trend-score-fill, eval=F}

model_trend <- read_csv(file.path(dir_goal, 'int/spp_trends_model.csv'))
spp_list    <- read_csv(file.path(dir_goal, 'int/all_spp_list.csv'))

# Add scores to each population trend:
pop_trend_score <- data.frame(pop_trend = c("Increasing", "Decreasing", "Stable", "Unknown"),
                              score = c(0.025, -0.025, 0, NA))

# Fill in NA trend scores with iucn population trend scores:
spp_trend_scores <- model_trend %>% 
  left_join(pop_trend_score, by = "pop_trend") %>% 
  filter(!is.na(pop_trend)) %>%  # Remove species with NAs
  mutate(trend_score = ifelse(is.na(trend_score), score, trend_score),
         trend_source = case_when(is.na(trend_source) & trend_score == score ~ "iucn_pop_trend",
                                  !is.na(trend_source) ~ trend_source)) %>% 
  left_join(spp_list, by = 'sciname') %>%  # Add class for averaging
  mutate(year = 2020,
         region_id = 1) %>% # Need to add a year and region_id for the toolbox to run
  dplyr::select(region_id, sciname, class, trend_score, trend_source, year) 

# Save to int folder 
write_csv(spp_trend_scores, file.path(dir_goal, 'int/spp_trend_scores.csv'))
```

## Finalize Trend Scores 

Use the same method as the current status to find the trend in species status. Trend will be calculated for each class and then the average of the trend for all the classes will be used as the overall trend.  

```{r final-trend-scores, eval=F}
spp_trend_scores <- read_csv(file.path(dir_goal, "int/spp_trend_scores.csv"))

spp_trend <- spp_trend_scores %>% 
  group_by(region_id, class) %>% # First find average score by class
  summarize(trend = mean(trend_score, na.rm=T)) %>%
  group_by(region_id) %>%
  summarize(trend = mean(trend, na.rm=T)) %>% # Find average of the 10 classes
  mutate( year = 2019) %>% 
  dplyr::select(region_id, year, trend_score = trend)

# Save to output folder
write_csv(spp_trend, file.path(dir_goal, "output/spp_trend.csv"))
```

## Save to Toolbox

```{r save-to-toolbox, eval=F}
trend_scores <- read_csv(file.path(dir_goal, 'output/spp_trend.csv'))
write_csv(trend_scores, file.path(dir_scores, 'layers/spp_trend.csv'))
```
